pyproject.toml
[project]
name = "churn-demo"
version = "0.1.0"
description = "Predict Churn"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml",
  "pydantic>=2.0",
]

[project.optional-dependencies]
# Training dependencies
training = [
  "google-cloud-bigquery",
  "pandas",
  "numpy<2",
  "xgboost",
  "optuna",
  "scikit-learn",
  "db-dtypes",
  "mlflow>=3.7.0",
  "torch>=2.2,<2.3",
  "google-cloud-storage",
]

# Inference/API dependencies
inference = [
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "mlflow>=3.7.0",
  "google-cloud-storage",
  "pandas",
  "numpy<2",
  "xgboost",
]

# Deployment dependencies
deploy = [
  "kfp",
  "google-cloud-aiplatform",
  "google-cloud-storage",
]

# Development dependencies
dev = [
  "pytest>=7.0",
  "pytest-cov>=4.0",
  "pre-commit",
  "joblib",
  "requests>=2.31.0",
]

[tool.black]
line-length = 100
target-version = ['py312']
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | \.terraform
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_glob = [".venv/*", "__pycache__/*", ".terraform/*"]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
]
